app.js 에 박혀있던 한글 처리로직 KoreanHealper.js 에 정리해서 옮김
